+++ account_document/models/__init__.py
@@ -9,3 +9,4 @@ from . import res_company
 from . import account_chart_template
 from . import account_payment_receiptbook
 from . import account_payment
+from . import account_reconciliation_widget
